requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community

# OpenAI Integration
langchain-openai
openai

# Environment Variable Management
python-dotenv

# pdf loader and parser lib
pypdf